requirements.txt
pytest
selenium
requests
pytest-xdist
allure-pytest
